pyproject.toml
[tool.poetry]
name = "rogalik"
version = "0.1.0"
description = ""
authors = ["XakepAnonim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.3"}
click = "^8.1.7"
python-socketio = "^5.11.4"
pytest = "^8.3.3"
ruff = "^0.7.0"
ruff-lsp = "^0.0.58"
pygame = "^2.6.1"
structure-generator-by-xakepanonim = "^0.1.1"
requests = "^2.32.3"
python-engineio = "^4.10.1"
websocket-client = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "pypackages",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "tests",
    "init.py",
    "**/migrations/*"
]
line-length = 79
indent-width = 4

target-version = "py313"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false

[tool.ruff.lint]
ignore = [
    "PLR0911", # Too many return statements (7 > 6)
    "C417", # Unnecessary map usage (rewrite using a generator expression)
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "RUF002", # check for cyrillic symbols
    "G004",  # Logging statement uses string formatting
    "D200",  # One-line docstring should fit on one line with quotes
    "D203",
    "D212",  # Multi-line docstring summary should start at the second line
    "D400", # First line should end with a period
    "D415", # First line should end with a period, question mark, or exclamation point
    "D205", # 1 blank line required between summary line and description
    "D107", # Missing docstring in init,
    "RUF003", # Comment contains ambiguous У (CYRILLIC CAPITAL LETTER U). Did you mean Y (LATIN CAPITAL LETTER Y)?
    "TID252", # Prefer absolute imports over relative imports from parent modules
    "RUF001", # check for latin symbols
    "PLR0913", # too-many-arguments
    "PERF401",
    "RUF012",
    "D106",
    "D100", # Missisng docstring in 1 line in .py files
    "D104", # Missing docstring in init
]

select = [
    "E" ,    # pycodestyle errors
    "W" ,    # pycodestyle warnings
    "F" ,    # pyflakes
    "B" ,    # flake8-bugbear
    "C4" ,   # flake8-comprehensions
    "UP" ,   # pyupgrade,
    "I" ,    # isort
    "YTT",   # flake8-2020
    "Q",     # flake8-quotes
    "T10",   # flake8-debugger
    "INT",   # flake8-gettext
    "PL",    # pylint
    "PT",    # flake8-pytest
    "PIE",   # misk-lints
    "PYI",   # flake8-pyi
    "TID",   # flake8-tidy-imports
    "ISC",   # implicit-str-concat
    "TCH",   # type checking imports
    "PGH",   # pygrep-helpers
    "RUF",   # ruff
    "S102",  # flake8-bandit
    "PERF",  # Perflint
    "FLY",   # flynt
    "G",     # flake8-logging-format
    "FA",    # flake8-annotations-complexity
    "D"
 ]

[tool.ruff.lint.per-file-ignores]
"/tests//*.py" = [
    # at least this three should be fine in tests:
    "S101", # asserts allowed in tests...
    "ARG", # Unused function args -> fixtures nevertheless are functionally relevant...
    "FBT", # Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
#     The below are debateable
    "PLR2004", # Magic value used in comparison, ...
    "S311", # Standard pseudo-random generators are not suitable for cryptographic purposes
]
"serializers.py" = ["D106"]
